requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
pandas>=1.3.0
reportlab>=3.5.67
psycopg2==2.9.10
easy-thumbnails ==2.9
django-filer == 3.3.1
WeasyPrint==61.2